Now matrix multiplication is done on the GPU if it would be faster.

+++ CSubNet/src/coreDefs.h
@@ -1,8 +1,8 @@
 #ifndef _CORE_DEFS_H
 #define _CORE_DEFS_H
 
-//typedef float netF;
-typedef double netF;
+//typedef double netF;
+typedef float netF;
 
 
 /*#ifdef __GNUC__
